+++ src/corelibs/U2Algorithm/src/util_msa_consensus/MSAConsensusAlgorithmDefault.cpp
@@ -49,7 +49,7 @@ MSAConsensusAlgorithm* MSAConsensusAlgorithmFactoryDefault::createAlgorithm(cons
 //////////////////////////////////////////////////////////////////////////
 // Algorithm
 
-char MSAConsensusAlgorithmDefault::getConsensusCharAndScore(const MultipleAlignment& msa, int pos, int& cnt, const QVector<qint64> &seqIdx) const {
+char MSAConsensusAlgorithmDefault::getConsensusCharAndScore(const MultipleAlignment& msa, int pos, int& cnt, const QVector<int> &seqIdx) const {
     //TODO: use var-length array!
     QVector<QPair<int, char> > freqs(32);
     int ch = U2Msa::GAP_CHAR;
